Make fixes according to #92 review comments
